rename python test files
