+++ onnxruntime/test/python/transformers/test_gqa_cuda.py
@@ -17,7 +17,6 @@ import unittest
 
 import numpy
 import torch
-from bert_padding import pad_input, unpad_input
 from einops import rearrange, repeat
 from onnx import TensorProto, helper
 from packaging import version
@@ -39,20 +38,18 @@ class Formats:
 class Config:
     batch_size = 0
     sequence_length = 0
-    kv_sequence_length = 0
-    past_sequence_length = 0
+    kv_sequence_length = 0  # this is past sequence length when there is past state.
     num_heads = 0
     kv_num_heads = 0
     head_size = 0
     ep = "CUDAExecutionProvider"
 
     def __init__(
-        self, batch_size, sequence_length, kv_sequence_length, past_sequence_length, num_heads, kv_num_heads, head_size
+        self, batch_size, sequence_length, kv_sequence_length, num_heads, kv_num_heads, head_size
     ):
         self.batch_size = batch_size
         self.sequence_length = sequence_length
         self.kv_sequence_length = kv_sequence_length
-        self.past_sequence_length = past_sequence_length
         self.num_heads = num_heads
         self.kv_num_heads = kv_num_heads
         self.head_size = head_size
@@ -61,7 +58,7 @@ class Config:
         short_ep = self.ep[: -len("ExecutionProvider")].lower()
         return (
             f"Config(batch_size={self.batch_size}, sequence_length={self.sequence_length}, "
-            f"kv_sequence_length={self.kv_sequence_length}, past_sequence_length={self.past_sequence_length}, "
+            f"kv_sequence_length={self.kv_sequence_length}, "
             f"num_heads={self.num_heads}, kv_num_heads={self.kv_num_heads}, head_size={self.head_size}, ep={short_ep})"
         )
 
@@ -101,118 +98,6 @@ class PromptConfig:
             f"kv_sequence_length={self.kv_sequence_length}, buffer_sequence_length={self.buffer_sequence_length}, "
             f"num_heads={self.num_heads}, kv_num_heads={self.kv_num_heads}, head_size={self.head_size}, ep={short_ep})"
         )
-
-
-def create_packed_multihead_attention_graph(config):
-    nodes = [
-        helper.make_node(
-            "PackedMultiHeadAttention",
-            [
-                "query",
-                "",
-                "",
-                "",
-                "token_offset",
-                "cumulative_sequence_length",
-            ],
-            ["output"],
-            "PackedMultiHeadAttention_0",
-            num_heads=config.num_heads,
-            domain="com.microsoft",
-        ),
-    ]
-
-    graph = helper.make_graph(
-        nodes,
-        "PackedMultiHeadAttention_Graph",
-        [
-            helper.make_tensor_value_info(
-                "query",
-                TensorProto.FLOAT16,
-                [
-                    -1,
-                    config.num_heads,
-                    3,
-                    config.head_size,
-                ],
-            ),
-            helper.make_tensor_value_info(
-                "token_offset", TensorProto.INT32, [config.batch_size, config.sequence_length]
-            ),
-            helper.make_tensor_value_info("cumulative_sequence_length", TensorProto.INT32, [config.batch_size + 1]),
-        ],
-        [
-            helper.make_tensor_value_info(
-                "output",
-                TensorProto.FLOAT16,
-                [-1, config.num_heads * config.head_size],
-            ),
-        ],
-    )
-
-    model = helper.make_model(graph)
-    return model.SerializeToString()
-
-
-def create_multihead_attention_graph(config):
-    nodes = [
-        helper.make_node(
-            "MultiHeadAttention",
-            [
-                "query",
-                "key",
-                "value",
-            ],
-            ["output"],
-            "MultiHeadAttention_0",
-            num_heads=config.num_heads,
-            domain="com.microsoft",
-        ),
-    ]
-
-    graph = helper.make_graph(
-        nodes,
-        "MultiHeadAttention_Graph",
-        [
-            helper.make_tensor_value_info(
-                "query",
-                TensorProto.FLOAT16,
-                [
-                    config.batch_size,
-                    config.sequence_length,
-                    config.num_heads * config.head_size,
-                ],
-            ),
-            helper.make_tensor_value_info(
-                "key",
-                TensorProto.FLOAT16,
-                [
-                    config.batch_size,
-                    config.kv_sequence_length,
-                    config.num_heads * config.head_size,
-                ],
-            ),
-            helper.make_tensor_value_info(
-                "value",
-                TensorProto.FLOAT16,
-                [
-                    config.batch_size,
-                    config.kv_sequence_length,
-                    config.num_heads * config.head_size,
-                ],
-            ),
-        ],
-        [
-            helper.make_tensor_value_info(
-                "output",
-                TensorProto.FLOAT16,
-                [config.batch_size, config.sequence_length, config.num_heads * config.head_size],
-            ),
-        ],
-    )
-
-    model = helper.make_model(graph)
-    return model.SerializeToString()
 
 
 def create_group_query_attention_graph_prompt(
@@ -575,204 +460,6 @@ def create_group_query_attention_graph_past(
     return model.SerializeToString()
 
 
-def generate_random_padding_mask(max_seqlen, batch_size, device, mode="random"):
-    assert mode in ["full", "random", "third"]
-    if mode == "full":
-        lengths = torch.full((batch_size, 1), max_seqlen, device=device, dtype=torch.int32)
-    elif mode == "random":
-        lengths = torch.randint(max(1, max_seqlen - 20), max_seqlen, (batch_size, 1), device=device)
-    else:
-        lengths = torch.randint(max_seqlen // 3, max_seqlen, (batch_size, 1), device=device)
-    padding_mask = repeat(torch.arange(max_seqlen, device=device), "s -> b s", b=batch_size) < lengths
-    return padding_mask
-
-
-def generate_qkv(q, k, v, query_padding_mask=None, key_padding_mask=None, kvpacked=False, qkvpacked=False):
-    """
-    Arguments:
-        q: (batch_size, seqlen_q, nheads, d)
-        k: (batch_size, seqlen_k, nheads_k, d)
-        v: (batch_size, seqlen_k, nheads_k, d)
-        query_padding_mask: (batch_size, seqlen), bool
-        key_padding_mask: (batch_size, seqlen), bool
-    """
-    assert not (kvpacked and qkvpacked)
-    batch_size, seqlen_q, nheads, d = q.shape
-    _, seqlen_k, nheads_k, _ = k.shape
-    assert k.shape == (batch_size, seqlen_k, nheads_k, d)
-    assert v.shape == (batch_size, seqlen_k, nheads_k, d)
-
-    if query_padding_mask is not None:
-        q_unpad, indices_q, cu_seqlens_q, max_seqlen_q = unpad_input(q, query_padding_mask)
-
-        def output_pad_fn(output_unpad):
-            return pad_input(output_unpad, indices_q, batch_size, seqlen_q)
-
-    else:
-        q_unpad = rearrange(q, "b s h d -> (b s) h d")
-        cu_seqlens_q = torch.arange(
-            0, (batch_size + 1) * seqlen_q, step=seqlen_q, dtype=torch.int32, device=q_unpad.device
-        )
-        max_seqlen_q = seqlen_q
-
-        def output_pad_fn(output_unpad):
-            return rearrange(output_unpad, "(b s) h d -> b s h d", b=batch_size)
-
-    if key_padding_mask is not None:
-        k_unpad, indices_k, cu_seqlens_k, max_seqlen_k = unpad_input(k, key_padding_mask)
-        v_unpad, _, _, _ = unpad_input(v, key_padding_mask)
-    else:
-        k_unpad = rearrange(k, "b s h d -> (b s) h d")
-        v_unpad = rearrange(v, "b s h d -> (b s) h d")
-        cu_seqlens_k = torch.arange(
-            0, (batch_size + 1) * seqlen_k, step=seqlen_k, dtype=torch.int32, device=k_unpad.device
-        )
-        max_seqlen_k = seqlen_k
-
-    if qkvpacked:
-        assert (query_padding_mask == key_padding_mask).all()
-        assert nheads == nheads_k
-        qkv_unpad = torch.stack([q_unpad, k_unpad, v_unpad], dim=1)
-        qkv = torch.stack([q, k, v], dim=2)
-        if query_padding_mask is not None:
-
-            def dqkv_pad_fn(dqkv_unpad):
-                return pad_input(dqkv_unpad, indices_q, batch_size, seqlen_q)
-
-        else:
-
-            def dqkv_pad_fn(dqkv_unpad):
-                return rearrange(dqkv_unpad, "(b s) t h d -> b s t h d", b=batch_size)
-
-        return (
-            qkv_unpad.detach().requires_grad_(),
-            cu_seqlens_q,
-            max_seqlen_q,
-            qkv.detach().requires_grad_(),
-            output_pad_fn,
-            dqkv_pad_fn,
-        )
-    elif kvpacked:
-        kv_unpad = torch.stack([k_unpad, v_unpad], dim=1)
-        kv = torch.stack([k, v], dim=2)
-        dq_pad_fn = output_pad_fn
-        if key_padding_mask is not None:
-
-            def dkv_pad_fn(dkv_unpad):
-                return pad_input(dkv_unpad, indices_k, batch_size, seqlen_k)
-
-        else:
-
-            def dkv_pad_fn(dkv_unpad):
-                return rearrange(dkv_unpad, "(b s) t h d -> b s t h d", b=batch_size)
-
-        return (
-            q_unpad.detach().requires_grad_(),
-            kv_unpad.detach().requires_grad_(),
-            cu_seqlens_q,
-            cu_seqlens_k,
-            max_seqlen_q,
-            max_seqlen_k,
-            q.detach().requires_grad_(),
-            kv.detach().requires_grad_(),
-            output_pad_fn,
-            dq_pad_fn,
-            dkv_pad_fn,
-        )
-    else:
-        dq_pad_fn = output_pad_fn
-        if key_padding_mask is not None:
-
-            def dk_pad_fn(dk_unpad):
-                return pad_input(dk_unpad, indices_k, batch_size, seqlen_k)
-
-        else:
-
-            def dk_pad_fn(dk_unpad):
-                return rearrange(dk_unpad, "(b s) h d -> b s h d", b=batch_size)
-
-        return (
-            q_unpad.detach().requires_grad_(),
-            k_unpad.detach().requires_grad_(),
-            v_unpad.detach().requires_grad_(),
-            cu_seqlens_q,
-            cu_seqlens_k,
-            max_seqlen_q,
-            max_seqlen_k,
-            q.detach().requires_grad_(),
-            k.detach().requires_grad_(),
-            v.detach().requires_grad_(),
-            output_pad_fn,
-            dq_pad_fn,
-            dk_pad_fn,
-        )
-
-
-def create_inputs(config: Config, kv_packed=False, qkv_packed=True):
-    qkv = torch.randn(
-        config.batch_size,
-        config.sequence_length,
-        3,
-        config.num_heads,
-        config.head_size,
-        device="cuda",
-        dtype=torch.float16,
-        requires_grad=False,
-    )
-    key_padding_mask = generate_random_padding_mask(
-        config.sequence_length, config.batch_size, device="cuda", mode="random"
-    )
-    qkv_unpad, cu_seqlens, max_seqlen, qkv, output_pad_fn, dqkv_pad_fn = generate_qkv(
-        *qkv.unbind(dim=2), key_padding_mask, key_padding_mask, kv_packed, qkv_packed
-    )
-    return qkv_unpad, cu_seqlens, max_seqlen, qkv, output_pad_fn, dqkv_pad_fn, key_padding_mask
-
-
-def generate_token_offset(cu_seqlens, max_seqlen):
-    token_offset = []
-    token_padset = []  # These are the indices that contain padding tokens
-    for i in range(1, len(cu_seqlens)):
-        start = i - 1
-        pre_seqlen = cu_seqlens[i - 1]
-        seqlen = cu_seqlens[i]
-        token_offset += range(start * max_seqlen, (start * max_seqlen) + (seqlen - pre_seqlen))
-        token_padset += range((start * max_seqlen) + (seqlen - pre_seqlen), i * max_seqlen)
-    return numpy.asarray(token_offset + token_padset, dtype=numpy.int32)
-
-
-def flash_attn_varlen_qkvpacked_func(qkv_unpad, cu_seqlens, token_offset, config, causal=False):
-    onnx_model_str = create_packed_multihead_attention_graph(config)
-    qkv_unpad = torch.swapdims(qkv_unpad, 1, 2)
-    ort_inputs = {
-        "query": qkv_unpad.detach().cpu().numpy(),
-        "token_offset": token_offset,
-        "cumulative_sequence_length": cu_seqlens.cpu().numpy(),
-    }
-    sess_options = SessionOptions()
-    ort_session = InferenceSession(onnx_model_str, sess_options, providers=[config.ep])
-    ort_output = ort_session.run(None, ort_inputs)
-    output = torch.tensor(ort_output)
-    return output
-
-
-def mha_func(q, k, v, config):
-    onnx_model_str = create_multihead_attention_graph(config)
-    q = torch.reshape(q, (config.batch_size, config.sequence_length, -1))
-    k = torch.reshape(k, (config.batch_size, config.kv_sequence_length, -1))
-    v = torch.reshape(v, (config.batch_size, config.kv_sequence_length, -1))
-    ort_inputs = {
-        "query": q.detach().cpu().numpy(),
-        "key": k.detach().cpu().numpy(),
-        "value": v.detach().cpu().numpy(),
-    }
-    sess_options = SessionOptions()
-    ort_session = InferenceSession(onnx_model_str, sess_options, providers=[config.ep])
-    ort_output = ort_session.run(None, ort_inputs)
-    ort_output = numpy.array(ort_output)
-    output = torch.tensor(ort_output)
-    return output
-
-
 def rotary_options_for_current_os():
     # Reference implementation of rotary uses triton, which is not available in Windows.
     # So we only test rotary in Linux right now.
@@ -1009,14 +696,6 @@ def gqa_past_func(
         return output, present_k, present_v
 
 
-def construct_causal_mask(seqlen_q, seqlen_k, query_padding_mask=None, key_padding_mask=None, device=None):
-    row_idx = rearrange(torch.arange(seqlen_q, device=device, dtype=torch.long), "s -> s 1")
-    col_idx = torch.arange(seqlen_k, device=device, dtype=torch.long)
-    sk = seqlen_k if key_padding_mask is None else rearrange(key_padding_mask.sum(-1), "b -> b 1 1 1")
-    sq = seqlen_q if query_padding_mask is None else rearrange(query_padding_mask.sum(-1), "b -> b 1 1 1")
-    return col_idx > row_idx + sk - sq
-
-
 def construct_local_mask(
     seqlen_q,
     seqlen_k,
@@ -1127,93 +806,6 @@ def attention_ref(
     return output.to(dtype=dtype_og), attention.to(dtype=dtype_og)
 
 
-def attention_qkvpacked_ref(
-    qkv,
-    key_padding_mask=None,
-    dropout_p=0.0,
-    dropout_mask=None,
-    causal=False,
-    upcast=True,
-    reorder_ops=False,
-    use_smooth_softmax=False,
-):
-    return attention_ref(
-        qkv[:, :, 0],
-        qkv[:, :, 1],
-        qkv[:, :, 2],
-        key_padding_mask,
-        key_padding_mask,
-        dropout_p,
-        dropout_mask,
-        upcast=upcast,
-        causal=causal,
-        reorder_ops=reorder_ops,
-        use_smooth_softmax=use_smooth_softmax,
-    )
-
-
-def parity_check_mha(
-    config,
-    packed,
-    rtol=1e-3,
-    atol=1e-3,
-):
-    if packed:
-        qkv_unpad, cu_seqlens, _, qkv, output_pad_fn, _, key_padding_mask = create_inputs(config)
-        token_offset = generate_token_offset(cu_seqlens, config.sequence_length).reshape(
-            (config.batch_size, config.sequence_length)
-        )
-        # ORT Flash
-        out_unpad = flash_attn_varlen_qkvpacked_func(qkv_unpad, cu_seqlens, token_offset, config, causal=False)
-        out_unpad = torch.squeeze(out_unpad, 0)
-        out = torch.reshape(
-            output_pad_fn(out_unpad), (config.batch_size, config.sequence_length, config.num_heads, config.head_size)
-        )
-        out = out.detach().cpu().numpy()
-        # Pytorch to compare
-        out_ref, _ = attention_qkvpacked_ref(qkv, key_padding_mask, 0.0, None, causal=False)
-        out_ref = out_ref.detach().cpu().numpy()
-    else:
-        q = torch.randn(
-            config.batch_size,
-            config.sequence_length,
-            config.num_heads,
-            config.head_size,
-            device="cuda",
-            dtype=torch.float16,
-            requires_grad=False,
-        )
-        k = torch.randn(
-            config.batch_size,
-            config.kv_sequence_length,
-            config.kv_num_heads,
-            config.head_size,
-            device="cuda",
-            dtype=torch.float16,
-            requires_grad=False,
-        )
-        v = torch.randn(
-            config.batch_size,
-            config.kv_sequence_length,
-            config.kv_num_heads,
-            config.head_size,
-            device="cuda",
-            dtype=torch.float16,
-            requires_grad=False,
-        )
-        out = mha_func(q, k, v, config)
-        out = torch.squeeze(out, 0)
-        out = torch.reshape(out, (config.batch_size, config.sequence_length, config.num_heads, config.head_size))
-        out = out.detach().cpu().numpy()
-        # Pytorch to compare
-        out_ref, _ = attention_ref(q, k, v, None, None, 0.0, None, causal=False)
-        out_ref = out_ref.detach().cpu().numpy()
-
-    numpy.testing.assert_allclose(
-        out, out_ref, rtol=rtol, atol=atol, equal_nan=True, err_msg=f" with {config} packed={packed}"
-    )
-
-
 def rotary_embedding(*args, **kwargs):
     # Use local import since triton is not available in Windows.
     from rotary_flash import apply_rotary_emb
@@ -1222,7 +814,7 @@ def rotary_embedding(*args, **kwargs):
 
 
 def parity_check_gqa_prompt(
-    config,
+    config: PromptConfig,
     causal=True,
     local=False,
     past_format=Formats.BNSH,
@@ -1420,7 +1012,7 @@ def parity_check_gqa_prompt(
 
 
 def parity_check_gqa_prompt_no_buff(
-    config,
+    config: PromptConfig,
     causal=True,
     local=False,
     past_format=Formats.BNSH,
@@ -1595,7 +1187,7 @@ def parity_check_gqa_prompt_no_buff(
 
 
 def parity_check_gqa_past(
-    config,
+    config: Config,
     causal=True,
     local=False,
     past_format=Formats.BNSH,
@@ -1788,7 +1380,7 @@ def parity_check_gqa_past(
 
 
 def parity_check_gqa_past_no_buff(
-    config,
+    config: Config,
     causal=True,
     local=False,
     past_format=Formats.BNSH,
@@ -2019,67 +1611,6 @@ def has_memory_efficient():
     return True
 
 
-def packed_mha_test_cases():
-    batches = [2] if pipeline_mode else [1, 5]
-    seqs = [1024, 1025] if pipeline_mode else [1024, 1025, 2048]
-    num_h = [1, 3] if pipeline_mode else [1, 6, 16]
-    h_sizes = [16, 256] if pipeline_mode else [32, 40, 64, 80, 96, 128, 160, 192, 224, 256]
-
-    for b in batches:
-        for s in seqs:
-            for n in num_h:
-                for h in h_sizes:
-                    config = Config(b, s, s, 0, n, n, h)
-                    yield str(config), config
-
-
-def mha_test_cases():
-    batches = [2] if pipeline_mode else [1, 5]
-    seqs = (
-        [(1, 128), (113, 211), (2048, 2048)]
-        if pipeline_mode
-        else [
-            (113, 203),
-            (128, 217),
-            (113, 211),
-            (108, 256),
-            (256, 512),
-            (512, 256),
-            (1024, 1024),
-            (1023, 1024),
-            (1024, 1023),
-            (2048, 2048),
-        ]
-    )
-    num_h = [3] if pipeline_mode else [1, 6, 16]
-    h_sizes = [64] if pipeline_mode else [32, 40, 64, 80, 96, 128, 160, 192, 224, 256]
-
-    for b in batches:
-        for s, s2 in seqs:
-            for n in num_h:
-                for h in h_sizes:
-                    config = Config(b, s, s2, 0, n, n, h)
-                    yield str(config), config
-
-
-class TestMHA(unittest.TestCase):
-    @parameterized.expand(packed_mha_test_cases())
-    def test_packed_mha(self, _, config):
-        if not has_flash_attention():
-            return
-        os.environ["ORT_DISABLE_FLASH_ATTENTION"] = "0"
-        print("-------- TEST PACKED MHA ---------")
-        parity_check_mha(config, True)
-
-    @parameterized.expand(mha_test_cases())
-    def test_mha(self, _, config):
-        if not has_flash_attention():
-            return
-        os.environ["ORT_DISABLE_FLASH_ATTENTION"] = "0"
-        print("-------- TEST MHA ---------")
-        parity_check_mha(config, False)
-
-
 def gqa_no_past_memory_efficient_test_cases():
     batches = [3] if pipeline_mode else [1, 3, 5]
     seqs = (
@@ -2196,9 +1727,7 @@ def gqa_past_memory_efficient_test_cases():
                                 for softcap in [0.0, 50.0]:
                                     if rotary and h % 16 > 0:
                                         continue
-
-                                    sp = random.randint(1, s2 - s) if s2 - s > 0 else 0
-                                    config = Config(b, s, s2, sp, n, n2, h)
+                                    config = Config(b, s, s2, n, n2, h)
                                     yield (
                                         str(config) + f"{local}_{rotary}_{rotary_interleaved}_{packed}_{softcap}",
                                         config,
@@ -2244,8 +1773,7 @@ def gqa_past_flash_attention_test_cases():
                                     if rotary and h % 16 > 0:
                                         continue
 
-                                    sp = random.randint(1, s2 - s) if s2 - s > 0 else 0
-                                    config = Config(b, s, s2, sp, n, n2, h)
+                                    config = Config(b, s, s2, n, n2, h)
                                     yield (
                                         str(config) + f"{local}_{rotary}_{rotary_interleaved}_{packed}_{softcap}",
                                         config,
@@ -2290,7 +1818,7 @@ def gqa_interactive_one_batch_flash_attention_test_cases():
                                 if rotary and h % 16 > 0:
                                     continue
 
-                                config = Config(b, s, s2, -1, n, n2, h)
+                                config = Config(b, s, s2, n, n2, h)
                                 yield (
                                     str(config) + f"{local}_{rotary}_{rotary_interleaved}_{packed}",
                                     config,
@@ -2333,7 +1861,7 @@ def gqa_interactive_one_batch_memory_efficient_attention_test_cases():
                             if rotary and h % 16 > 0:
                                 continue
 
-                            config = Config(b, s, s2, -1, n, n2, h)
+                            config = Config(b, s, s2, n, n2, h)
                             yield (
                                 str(config) + f"{rotary}_{rotary_interleaved}_{packed}",
                                 config,
